issue-29 Change end to end test to be an integration test for OPEN command
